fix(desktop): 修复跨平台编译和 helper 二进制独立构建
问题：helper 二进制编译失败，因为 lib.rs 无条件依赖 tauri crate，
且 tauri-build 在所有目标构建时运行。

修复：
- lib.rs: 拆分为模块声明 + #[cfg(feature = desktop)] include!
  tauri 相关代码移到 lib_tauri.rs，helper 编译时跳过
- Cargo.toml: tauri-build 改为 optional，仅在 desktop feature 启用
- build.rs: #[cfg(feature = tauri-build)] 条件执行
- remote_commands.rs: 移除 tokio::spawn_blocking 依赖，
  所有远程命令改为 sync fn，由 Tauri 运行时自动分派到线程池
- fs_ext.rs: Windows Perissions::from_mode 使用 metadata 获取权限
- proc.rs: #[cfg(unix)] 守卫 sh 相关测试
- 修复 unused variable 和 import warnings

验证：
- cargo check --bin csswitch-helper --no-default-features ✅
- cargo check --lib ✅
- cargo test --lib: 31 passed, 0 failed ✅
